add copyTableAndUpdate against props
