+++ src/components/side-panel/layer-manager.js
@@ -136,7 +136,6 @@ function LayerManagerFactory(AddDataButton, LayerPanel, SourceDataCatalog) {
   class LayerManager extends Component {
     static propTypes = {
       datasets: PropTypes.object.isRequired,
-      datasetUpdateNumbers: PropTypes.object.isRequired,
       layerBlending: PropTypes.string.isRequired,
       layerClasses: PropTypes.object.isRequired,
       layers: PropTypes.arrayOf(PropTypes.any).isRequired,
@@ -191,7 +190,6 @@ function LayerManagerFactory(AddDataButton, LayerPanel, SourceDataCatalog) {
       const {
         layers,
         datasets,
-        datasetUpdateNumbers,
         intl,
         layerOrder,
         showAddDataModal,
@@ -218,7 +216,6 @@ function LayerManagerFactory(AddDataButton, LayerPanel, SourceDataCatalog) {
 
       const panelProps = {
         datasets,
-        datasetUpdateNumbers,
         openModal,
         layerTypeOptions
       };
@@ -227,7 +224,6 @@ function LayerManagerFactory(AddDataButton, LayerPanel, SourceDataCatalog) {
         <div className="layer-manager">
           <SourceDataCatalog
             datasets={datasets}
-            datasetUpdateNumbers={datasetUpdateNumbers}
             showDatasetTable={showDatasetTable}
             updateDatasetColor={updateDatasetColor}
             removeDataset={removeDataset}
