+++ src/utils/index.d.ts
@@ -2,4 +2,4 @@
 export {maybeToDate, roundValToStep} from './data-utils';
 export {updateAllLayerDomainData} from '../reducers/vis-state-updaters';
 // export {getHexFields} from '../layers/h3-hexagon-layer/h3-utils';
-export {default as KeplerTable, Field, findPointFieldPairs} from './table-utils/kepler-table';
+export {default as KeplerTable, Field, findPointFieldPairs, copyTableAndUpdate} from './table-utils/kepler-table';
