+++ test/browser/components/side-panel/filter-manager-test.js
@@ -56,7 +56,6 @@ const nop = () => {};
 const defaultProps = {
   filters: [],
   datasets: {},
-  datasetUpdateNumbers: {},
   layers: [],
   showDatasetTable: nop,
   visStateActions: {
@@ -73,7 +72,6 @@ const defaultProps = {
 const filterManagerProps = {
   filters: StateWFilters.visState.filters,
   datasets: StateWFilters.visState.datasets,
-  datasetUpdateNumbers: StateWFilters.visState.datasetUpdateNumbers,
   layers: StateWFilters.visState.layers,
   showDatasetTable: nop,
   visStateActions: {
